requirements.txt
pytz
dotenv
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
jinja2
google-genai
pydantic
tenacity